fixed sub key not existing on sub report page

+++ src/Report/Formats/HTML.php
@@ -48,7 +48,7 @@ class HTML implements Format
 
             if (empty($this->printTheseArrayKeys)) {
                 foreach ($row as $item => $val) {
-					if($item == "conversions" && $val > 0 && $row["sub"] != "TOTAL") {
+					if($item == "conversions" && $val > 0 && (key_exists('sub', $row) && $row["sub"] != "TOTAL") ) {
 						echo "<td><a href='/report/sub/conversions?subid={$row["sub"]}". "&" . "{$params}'>{$val}</a></td>";
 					} else {
 						echo "<td>{$val}</td>";
